# openapi/deposit-recovery.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deposit Recovery API",
    "description": "Qull Connect API. Supply a per-user API key in Authorization: Bearer <key>. Keys identify one customer; arbitrary platform/user-id headers are not accepted. Payment setup opens hosted Stripe Checkout and does not charge a service fee. Use the record's billing/status operation to verify setup, then explicitly authorize the applicable fee. See the linked guide for workflow and service limits.",
    "version": "1.0.0",
    "contact": {
      "name": "Qull, Inc.",
      "email": "wasiq@qull.io"
    }
  },
  "servers": [
    {
      "url": "https://5.78.152.6.nip.io/deposit-recovery"
    }
  ],
  "paths": {
...
    "/api/life-events": {
      "post": {
        "summary": "Api Life Events",
        "description": "Receive a fan-out life event from the shared bus.\n\nThe caller's verified API key determines ownership. Client-supplied\nowner IDs and legacy records without an owner are never claimable.\nCreates a real case when the payload is complete, otherwise a draft case\npre-filled with everything the payload provides. Always returns the\nproactive nudge in `user_message`.",
        "operationId": "api_life_events_api_life_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LifeEventIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or disabled per-user API key."
          },
          "413": {
            "description": "Request body exceeds the configured limit."
          },
          "429": {
            "description": "Request rate limit exceeded. Wait for the Retry-After interval."
          }
        }
      }
    },
...
    "/api/me/data": {
      "get": {
        "summary": "Export My Data",
        "description": "Export only the authenticated caller's local records.",
        "operationId": "export_my_data_api_me_data_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or disabled per-user API key."
          },
          "413": {
            "description": "Request body exceeds the configured limit."
          },
          "429": {
            "description": "Request rate limit exceeded. Wait for the Retry-After interval."
          }
        }
      },
      "delete": {
        "summary": "Delete My Data",
        "description": "Delete local records and documents. Stripe/payment audit records remain separately retained.",
        "operationId": "delete_my_data_api_me_data_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteDataIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or disabled per-user API key."
          },
          "413": {
            "description": "Request body exceeds the configured limit."
          },
          "429": {
            "description": "Request rate limit exceeded. Wait for the Retry-After interval."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteDataIn": {
        "properties": {
          "confirm_delete": {
            "type": "boolean",
            "const": true,
            "title": "Confirm Delete",
            "description": "The user explicitly confirms deleting their local records and generated documents."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "confirm_delete"
        ],
        "title": "DeleteDataIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LifeEventIn": {
        "properties": {
          "event_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Event Type"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_type"
        ],
        "title": "LifeEventIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "UserApiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Qull per-user API key",
        "description": "An operator-provisioned opaque API key belonging to one customer. Never use a Stripe secret key here."
      }
    }
  },
  "security": [
    {
      "UserApiKey": []
    }
  ]
}